Moved sleep timer to a flyout.
