+++ src/Neptunium/App.xaml.cs
@@ -68,7 +68,6 @@ namespace Neptunium
         internal static void RegisterUIDialogs()
         {
             NepApp.UI.Overlay.RegisterDialogFragment<StationInfoDialogFragment, StationInfoDialog>();
-            NepApp.UI.Overlay.RegisterDialogFragment<SleepTimerDialogFragment, SleepTimerDialog>();
         }
 
         private static volatile bool isInBackground = false;
