+++ src/Neptunium/Neptunium.csproj
@@ -157,7 +157,7 @@
     <Compile Include="MediaExtensions.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="ViewModel\AppShellViewModel.cs" />
-    <Compile Include="ViewModel\Dialog\SleepTimerDialogFragment.cs" />
+    <Compile Include="ViewModel\Fragments\SleepTimerContextFragment.cs" />
     <Compile Include="ViewModel\Dialog\StationInfoDialogFragment.cs" />
     <Compile Include="ViewModel\NowPlayingPageViewModel.cs" />
     <Compile Include="ViewModel\SettingsPageViewModel.cs" />
@@ -168,9 +168,6 @@
     </Compile>
     <Compile Include="Core\UI\ProgressIndicatorDialog.xaml.cs">
       <DependentUpon>ProgressIndicatorDialog.xaml</DependentUpon>
-    </Compile>
-    <Compile Include="View\Dialog\SleepTimerDialog.xaml.cs">
-      <DependentUpon>SleepTimerDialog.xaml</DependentUpon>
     </Compile>
     <Compile Include="View\Dialog\StationInfoDialog.xaml.cs">
       <DependentUpon>StationInfoDialog.xaml</DependentUpon>
@@ -331,10 +328,6 @@
       <SubType>Designer</SubType>
       <Generator>MSBuild:Compile</Generator>
     </Page>
-    <Page Include="View\Dialog\SleepTimerDialog.xaml">
-      <SubType>Designer</SubType>
-      <Generator>MSBuild:Compile</Generator>
-    </Page>
     <Page Include="View\Dialog\StationInfoDialog.xaml">
       <SubType>Designer</SubType>
       <Generator>MSBuild:Compile</Generator>
@@ -368,9 +361,7 @@
       <Generator>MSBuild:Compile</Generator>
     </Page>
   </ItemGroup>
-  <ItemGroup>
-    <Folder Include="ViewModel\Fragments\" />
-  </ItemGroup>
+  <ItemGroup />
   <ItemGroup>
     <SDKReference Include="Microsoft.VCLibs, Version=14.0">
       <Name>Visual C++ 2015 Runtime for Universal Windows Platform Apps</Name>
