--- src/Neptunium/View/Dialog/SleepTimerDialog.xaml
@@ -1,54 +1,0 @@
-﻿<Page
-    x:Class="Neptunium.View.Dialog.SleepTimerDialog"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:Neptunium.View.Dialog"
-    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    mc:Ignorable="d">
-
-    <Grid Background="{ThemeResource ApplicationPageBackgroundThemeBrush}">
-        <Grid.RowDefinitions>
-            <RowDefinition Height="50" />
-            <RowDefinition Height="65" />
-            <RowDefinition Height="*" />
-            <RowDefinition Height="75" />
-        </Grid.RowDefinitions>
-
-        <Grid Grid.Row="0" Margin="5,5,5,5">
-            <TextBlock Margin="10 0" Text="Sleep Timer"
-                       Foreground="{ThemeResource SystemControlForegroundBaseHighBrush}"
-                       Style="{ThemeResource TitleTextBlockStyle}" />
-        </Grid>
-
-        <ComboBox Grid.Row="1" Margin="15 0" Header="Time to wait:" 
-                  HorizontalAlignment="Stretch"
-                  ItemsSource="{Binding AvailableSleepItems, UpdateSourceTrigger=PropertyChanged}"
-                  SelectedItem="{Binding SelectedSleepItem, Mode=TwoWay, UpdateSourceTrigger=PropertyChanged}"
-                  DisplayMemberPath="DisplayName">
-        </ComboBox>
-
-        <TextBlock Grid.Row="2" Margin="15 0"
-                   Style="{ThemeResource SubtitleTextBlockStyle}"
-                   Foreground="{ThemeResource SystemControlForegroundBaseMediumHighBrush}">
-                <Span>
-                    <Run Text="Estimated Sleep Time:" />
-                    
-                    <Run Text="{Binding EstimatedTime, UpdateSourceTrigger=PropertyChanged}" 
-                         Foreground="{ThemeResource SystemControlForegroundAccentBrush}" />
-                </Span>
-        </TextBlock>
-
-        <StackPanel Grid.Row="3" Orientation="Horizontal" HorizontalAlignment="Right">
-            <StackPanel.Resources>
-                <Style TargetType="Button">
-                    <Setter Property="Margin" Value="0 0 10 0" />
-                </Style>
-            </StackPanel.Resources>
-            <Button Content="Cancel" Grid.Column="0" Command="{Binding CancelCommand}" MinWidth="75" x:Name="CancelButton" />
-
-            <Button Content="OK" Grid.Column="1" Command="{Binding OKCommand}"
-                    Background="{ThemeResource AccentButtonBackground}" MinWidth="75" />
-        </StackPanel>
-    </Grid>
-</Page>
